# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CMS Backend API",
    "version": "1.0.0",
    "description": "API for managing CMS pages and menu types"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "paths": {
...
    "/cms-pages/file/{id}": {
      "get": {
        "summary": "Download/View Page File",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "File retrieved"
          }
        }
      }
    }
  },
...
}